compose elementwise operations

+++ utils.h
@@ -12,7 +12,7 @@ void apply_relu(Float* Begin, float* End)
 {
     for (; Begin != End; ++Begin)
     {
-        *Begin = std::max(0.f, *Begin);
+        *Begin = std::max(static_cast<Float>(0), *Begin);
     }
 }
